refactor: move rejected exam event handlers
